+++ CHANGELOG.md
@@ -44,6 +44,8 @@ Version 1.5.1 (not released yet)
   Windows versions because of the lack of support for WinHelp format
 - fix F4 key (Maximize active viewport) in RED for screens larger than 1024x768
 - fix shooting through transparent sections of DDS textures
+- change default server update rate to 20
+- add 'update_rate' command server-side and client-side
 
 Version 1.5.0 (released 2020-09-05)
 -----------------------------------